Added support for a private URL
In some cases, a bot may need to access the API using a different URL
than normal users would be accessing, such as using an SSL connection or
including HTTP authentication data in the URL for SOAPpy.  This patch
adds support for a private URL that is used by the bot to connect to the
API, while still using the normal URL when displaying bug data to users.

This adds a new configuration option 'plugins.Mantis.privateurlbase'.